Generalized input element for Listable element variables
Before a prefix was assumed but no longer.
Also the variable name was assumed to be the id of the input variable. Now it is first checked against the id, then checked against the name of fields.
Finally inputs are searched for within the context of the form id. In the future form id will be a selector probably.